+++ pyquotex/webapi/routers/__init__.py
@@ -1,0 +1,1 @@
+"""FastAPI routers for the pyquotex web API."""